Locl training limits. Start experimenting with SageMaker fitting job
