+++ train/model.py
@@ -1,35 +1,35 @@
+import torch
 import torch.nn as nn
+from torch.autograd import Variable
 
-class RNN(nn.Module):
-    def __init__(self, input_size, output_size, hidden_dim, n_layers):
-        super(RNN, self).__init__()
-        
+class LSTM(nn.Module):
+
+    def __init__(self, input_dim, hidden_dim, batch_size, output_dim=1,num_layers=2):
+        super(LSTM, self).__init__()
+        self.input_dim = input_dim
         self.hidden_dim = hidden_dim
-        self.n_layers = n_layers
+        self.batch_size = batch_size
+        self.num_layers = num_layers
 
-        # define an RNN with specified parameters
-        # batch_first means that the first dim of the input and output will be the batch_size
-        self.rnn = nn.RNN(input_size, hidden_dim, n_layers, batch_first=True)
+        # Define the LSTM layer
+        self.lstm = nn.LSTM(self.input_dim, self.hidden_dim, self.num_layers, dropout=0.25)
+
+        # Define the output layer
+        self.linear = nn.Linear(self.hidden_dim, output_dim)
+
+    def init_hidden(self):
+        # This is what we'll initialise our hidden state as
+        return (Variable(torch.zeros(self.num_layers, self.batch_size, self.hidden_dim)),
+                Variable(torch.zeros(self.num_layers, self.batch_size, self.hidden_dim)))
+
+    def forward(self, input):
+        # Forward pass through LSTM layer
+        # shape of lstm_out: [input_size, batch_size, hidden_dim]
+        # shape of self.hidden: (a, b), where a and b both 
+        # have shape (num_layers, batch_size, hidden_dim).
+        lstm_out, self.hidden1 = self.lstm(input.view(len(input), self.batch_size, -1))
         
-        # last, fully-connected layer
-        self.fc = nn.Linear(hidden_dim, output_size)
-        
-        
-    def forward(self, x, hidden):
-        # x (batch_size, seq_length, input_size)
-        # hidden (n_layers, batch_size, hidden_dim)
-        # r_out (batch_size, time_step, hidden_size)
-        batch_size = x.size(0)
-        
-        # get RNN outputs
-        r_out, hidden_out = self.rnn(x, hidden)
-        
-        # shape output to be (batch_size*seq_length, hidden_dim)
-        r_out = r_out.view(-1, self.hidden_dim)  
-        
-        # get final output 
-        output = self.fc(r_out)
-        # get last batch
-        output = output[:, -1]
-        
-        return output, hidden
+        # Only take the output from the final timestep
+        # Can pass on the entirety of lstm_out to the next layer if it is a seq2seq prediction
+        y_pred = self.linear(lstm_out[-1].view(self.batch_size, -1))
+        return y_pred.view(-1)
