--- train/modelLstm.py
@@ -1,37 +1,0 @@
-import torch
-import torch.nn as nn
-from torch.autograd import Variable
-
-class LSTM(nn.Module):
-
-    def __init__(self, input_dim, hidden_dim, batch_size, output_dim=1,num_layers=2):
-        super(LSTM, self).__init__()
-        self.input_dim = input_dim
-        self.hidden_dim = hidden_dim
-        self.batch_size = batch_size
-        self.num_layers = num_layers
-
-        # Define the LSTM layer
-        self.lstm = nn.LSTM(self.input_dim, self.hidden_dim, self.num_layers)
-
-        # Define the output layer
-        self.linear = nn.Linear(self.hidden_dim, output_dim)
-
-    def init_hidden(self):
-        # This is what we'll initialise our hidden state as
-        return (Variable(torch.zeros(self.num_layers, self.batch_size, self.hidden_dim)),
-                Variable(torch.zeros(self.num_layers, self.batch_size, self.hidden_dim)))
-
-    def forward(self, input):
-        # Forward pass through LSTM layer
-        # shape of lstm_out: [input_size, batch_size, hidden_dim]
-        # shape of self.hidden: (a, b), where a and b both 
-        # have shape (num_layers, batch_size, hidden_dim).
-        lstm_out, self.hidden = self.lstm(input.view(len(input), self.batch_size, -1))
-        
-        # Only take the output from the final timestep
-        # Can pass on the entirety of lstm_out to the next layer if it is a seq2seq prediction
-        y_pred = self.linear(lstm_out[-1].view(self.batch_size, -1))
-        return y_pred.view(-1)
-
-# model = LSTM(lstm_input_size, h1, batch_size=num_train, output_dim=output_dim, num_layers=num_layers)
