+++ train/train.py
@@ -9,7 +9,7 @@ import torch
 import torch.optim as optim
 import torch.utils.data
 
-from model import RNN
+from model import LSTM
 
 def model_fn(model_dir):
     """Load the PyTorch model from the `model_dir` directory."""
@@ -25,12 +25,17 @@ def model_fn(model_dir):
 
     # Determine the device and construct the model.
     device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
-    model = LSTMClassifier(model_info['embedding_dim'], model_info['hidden_dim'], model_info['vocab_size'])
+    model = LSTM(input_dim=2, 
+             hidden_dim=model_info['hidden_dim'], 
+             batch_size=model_info['batch_size'],
+             output_dim=1, 
+             num_layers=model_info['num_layers']).to(device)
 
     # Load the stored model parameters.
     model_path = os.path.join(model_dir, 'model.pth')
     with open(model_path, 'rb') as f:
         model.load_state_dict(torch.load(f))
+
 
     model.to(device).eval()
 
@@ -39,50 +44,61 @@ def model_fn(model_dir):
 
 def _get_train_data_loader(batch_size, training_dir):
     print("Get train data loader.")
+       
+    dummy_read = pd.read_csv(os.path.join(training_dir, 'train.csv'), header=None, names=None)
+    num_rows_train = (len(dummy_read) // batch_size) * batch_size
+    train_data = pd.read_csv(os.path.join(training_dir, 'train.csv'), header=None, names=None, nrows=num_rows_train)
 
-    train_data = pd.read_csv(os.path.join(training_dir, "train.csv"), header=None, names=None)
-    # Turn the input pandas dataframe into tensors
-    # For Y, we select 3rd column - average price next day
-    # For X, we select all but first 3 columns - all data for the current day
-    train_y = torch.from_numpy(train_data[[2]].values).float().squeeze()
-    train_X = torch.from_numpy(train_data.drop([0,1,2], axis=1).values).float()
+    train_y = torch.from_numpy(train_data[[0]].values).float().squeeze()
+    train_X = torch.from_numpy(train_data.drop([0], axis=1).values).long()
 
     train_ds = torch.utils.data.TensorDataset(train_X, train_y)
 
     return torch.utils.data.DataLoader(train_ds, batch_size=batch_size)
 
 
+def train(model, train_loader, epochs, optimizer, loss_fn, device):
+    """
+    This is the training method that is called by the PyTorch training script. The parameters
+    passed are as follows:
+    model        - The PyTorch model that we wish to train.
+    train_loader - The PyTorch DataLoader that should be used during training.
+    epochs       - The total number of epochs to train for.
+    optimizer    - The optimizer to use during training.
+    loss_fn      - The loss function used for training.
+    device       - Where the model and data should be loaded (gpu or cpu).
+    """
+        
+    # for dev
+    print_every = 5
+    batch_size=50
+    clip=5 # gradient clipping
+    criterion = loss_fn
 
-def train(model, train_loader, epochs, optimizer, criterion, device):
-    
-    # initialize the hidden state
-    hidden = None
-    
     for epoch in range(1, epochs + 1):
         model.train()
         total_loss = 0
-        
-        
         for batch in train_loader:
-            batch_X, batch_y = batch            
-            batch_X = batch_X.to(device)
-            batch_y = batch_y.to(device)
+            batch_X, batch_y = batch
             
             # convert data into Tensors
-            # x_tensor = torch.Tensor(x).unsqueeze(0) # unsqueeze gives a 1, batch_size dimension
-            # y_tensor = torch.Tensor(y)
             x_tensor = torch.Tensor(batch_X).unsqueeze(0)
+            y_tensor = torch.Tensor(batch_y)
+          
+            x_tensor = x_tensor.to(device)
+            y_tensor = y_tensor.to(device)
             
             model.zero_grad()
-            output, hidden = model(x_tensor, hidden)
             
-            loss = criterion(output.squeeze(), batch_y.float())
-            loss.backward(retain_graph=True)
-            optimizer.step()
+            output = model(x_tensor)
+           
+            loss = criterion(output.squeeze(), y_tensor)
+            loss.backward()
+            nn.utils.clip_grad_norm_(model.parameters(), clip)
+            optimizer.step()   
             
             total_loss += loss.data.item()
         print("Epoch: {}, MSELoss: {}".format(epoch, total_loss / len(train_loader)))
-    return model
 
 
 if __name__ == '__main__':
@@ -92,20 +108,20 @@ if __name__ == '__main__':
     parser = argparse.ArgumentParser()
 
     # Training Parameters
-    parser.add_argument('--batch-size', type=int, default=512, metavar='N',
+    parser.add_argument('--batch-size', type=int, default=50, metavar='N',
                         help='input batch size for training (default: 512)')
-    parser.add_argument('--epochs', type=int, default=10, metavar='N',
+    parser.add_argument('--epochs', type=int, default=100, metavar='N',
                         help='number of epochs to train (default: 10)')
     parser.add_argument('--seed', type=int, default=1, metavar='S',
                         help='random seed (default: 1)')
 
     # Model Parameters
-    parser.add_argument('--embedding_dim', type=int, default=32, metavar='N',
-                        help='size of the word embeddings (default: 32)')
-    parser.add_argument('--hidden_dim', type=int, default=100, metavar='N',
+    parser.add_argument('--hidden-dim', type=int, default=100, metavar='N',
                         help='size of the hidden dimension (default: 100)')
-    parser.add_argument('--vocab_size', type=int, default=5000, metavar='N',
-                        help='size of the vocabulary (default: 5000)')
+    parser.add_argument('--num-layers', type=int, default=4, metavar='N',
+                        help='number of LSTM layers (default: 4)')
+    parser.add_argument('--learning-rate', type=float, default=0.0001, metavar='N',
+                        help='learnign rate parameter (default: 0.0001)')
 
     # SageMaker Parameters
     parser.add_argument('--hosts', type=list, default=json.loads(os.environ['SM_HOSTS']))
@@ -125,18 +141,19 @@ if __name__ == '__main__':
     train_loader = _get_train_data_loader(args.batch_size, args.data_dir)
 
     # Build the model.
-    model = LSTMClassifier(args.embedding_dim, args.hidden_dim, args.vocab_size).to(device)
+    model = LSTM(input_dim=2, 
+                 hidden_dim=args.hidden_dim, 
+                 batch_size=args.batch_size,
+                 output_dim=1, 
+                 num_layers=args.num_layers).to(device)
 
-    with open(os.path.join(args.data_dir, "word_dict.pkl"), "rb") as f:
-        model.word_dict = pickle.load(f)
-
-    print("Model loaded with embedding_dim {}, hidden_dim {}, vocab_size {}.".format(
-        args.embedding_dim, args.hidden_dim, args.vocab_size
+    print("Model loaded with hidden_dim {}, batch_size {}, num_layers {}.".format(
+        args.hidden_dim, args.batch_size, args.num_layers
     ))
 
     # Train the model.
-    optimizer = optim.Adam(model.parameters())
-    loss_fn = torch.nn.BCELoss()
+    optimizer = optim.Adam(model.parameters(), lr=args.learning_rate)
+    loss_fn = torch.nn.MSELoss()
 
     train(model, train_loader, args.epochs, optimizer, loss_fn, device)
 
@@ -144,9 +161,9 @@ if __name__ == '__main__':
     model_info_path = os.path.join(args.model_dir, 'model_info.pth')
     with open(model_info_path, 'wb') as f:
         model_info = {
-            'embedding_dim': args.embedding_dim,
             'hidden_dim': args.hidden_dim,
-            'vocab_size': args.vocab_size,
+            'batch_size': args.batch_size,
+            'num_layers': args.num_layers,
         }
         torch.save(model_info, f)
